Safari-609.1.11.1/LayoutTests/fast/viewport/ios/viewport-shrink-to-fit-on-new-navigation-expected.txt

ALERT: pageshow - not from cache
ALERT: pageshow - from cache
window size: [1008, 1726]
zoom scale: 0.32
